Implemented BFS Search
The methods being used previously had issues with
time complexity. The new approach is to imlement
concepts of graph theory where each node in the
network is the id of the user and each node has
children where the children are the connections
that are formed between the target user and all
users that reference the original user id. By
Implmenting a Breadth First Search (BFS) the
code base can be refactored to use this to solve
the list of problems in the orignal prompt.